+++ src/rockverse/_utils/__init__.py
@@ -27,6 +27,7 @@ def rvtqdm(*args, **kwargs):
     return tqdm(*args, **kwargs)
 
 
+
 # decorator that copies docstrings
 def copy_docstring(func_to_copy_from):
     def decorator(func):